src/connectors/whatsapp/whatsapp_api.py
# Standard library import
from src.logger import logging

# Third-party imports
from twilio.rest import Client
from decouple import config


# Find your Account SID and Auth Token at twilio.com/console
# and set the environment variables. See http://twil.io/secure
account_sid = config("TWILIO_ACCOUNT_SID")
auth_token = config("TWILIO_AUTH_TOKEN")
client = Client(account_sid, auth_token)
twilio_number = config('TWILIO_NUMBER')

# Sending message logic through Twilio Messaging API
def send_message(to_number, body_text):
    try:
        message = client.messages.create(
            from_=f"whatsapp:{twilio_number}",
            body=body_text,
            to=f"whatsapp:{to_number}"
            )
        logging.info(f"Message sent to {to_number}: {message.body}")
    except Exception as e:
        logging.error(f"Error sending message to {to_number}: {e}")

if __name__ == "__main__":
    send_message(to_number="+919175373448", body_text="heello")