requirements.txt
python-dotenv
twilio
openai
python-decouple
flask